user_manager/requirements.txt
grpcio
grpcio-tools
flask
mysql-connector-python